+++ webui/cypress/integration/blacklist-pages.spec.js
@@ -86,7 +86,7 @@ describe('Blacklist pages', () => {
         });
       })
       .then(() => {
-        cy.get('a[href="/abc/admin/test/repair"]').should('exist').click();
+        cy.get('a:contains(Repair Queues)').should('exist').click();
         cy.get('a[href="/abc/admin/test/repair/input"]').should('exist');
         cy.get('a[href="/abc/admin/test/repair/output"]').should('exist');
         cy.get('a[href="/abc/admin/test/repair/jobs"]').should('not.exist');