+++ webui/src/pages/Cluster/components/IssuesButton/components/IconOkContrast/IconOkContrast.tsx
@@ -2,7 +2,7 @@
 import React, { memo } from 'react';
 import { css, cx } from '@emotion/css';
 // @ts-ignore
-import { IconOk } from '@tarantool.io/ui-kit';
+import { GenericIconProps, IconOk } from '@tarantool.io/ui-kit';
 
 const styles = {
   contrastIcon: css`
@@ -10,6 +10,8 @@ const styles = {
   `,
 };
 
-const IconOkContrast = ({ className, props }) => <IconOk className={cx(styles.contrastIcon, className)} {...props} />;
+const IconOkContrast = ({ className, ...props }: GenericIconProps) => (
+  <IconOk className={cx(styles.contrastIcon, className)} {...props} />
+);
 
 export default memo(IconOkContrast);